Major Bugs Fixed

+++ lib/ui/views/chat_list/chat_list_viewmodel.dart
@@ -1,12 +1,13 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_feather_icons/flutter_feather_icons.dart';
 import 'package:hint/api/firestore.dart';
 import 'package:hint/api/hive.dart';
 import 'package:hint/constants/app_keys.dart';
 import 'package:hint/constants/app_strings.dart';
 import 'package:hint/services/auth_service.dart';
 import 'package:hint/services/chat_service.dart';
+import 'package:hint/ui/shared/alert_dialog.dart';
 import 'package:hint/ui/shared/ui_helpers.dart';
 import 'package:stacked/stacked.dart';
 import 'package:hint/app/app_logger.dart';
@@ -25,19 +26,22 @@ class ChatListViewModel extends StreamViewModel<QuerySnapshot> {
 
   Future<void> showTileOptions(String fireuserId, context, bool pinned) async {
     return showModalBottomSheet(
+      shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(32)),
       context: context,
       builder: (context) {
-        return SizedBox(
-          height: screenHeightPercentage(context, percentage: 0.3),
+        return Padding(
+          padding: const EdgeInsets.symmetric(horizontal: 20),
           child: Column(
+            mainAxisSize: MainAxisSize.min,
             children: [
+              verticalSpaceMedium,
               ListTile(
                 onTap: () {
                   chatService.addToArchive(fireuserId);
                   hiveApi.addToArchivedUsers(fireuserId);
                   Navigator.pop(context);
                 },
-                leading: const Icon(CupertinoIcons.archivebox_fill),
+                leading: const Icon(FeatherIcons.folderPlus),
                 title: Text('Archive',
                     style: Theme.of(context)
                         .textTheme
@@ -55,10 +59,14 @@ class ChatListViewModel extends StreamViewModel<QuerySnapshot> {
                       : hiveApi.addToPinnedUsers(fireuserId);
                   Navigator.pop(context);
                 },
-                leading: Icon(pinned
-                    ? CupertinoIcons.pin_slash_fill
-                    : CupertinoIcons.pin_fill),
-                title: Text(pinned ? 'Remove Pinned' : 'Pinned',
+                leading: Padding(
+                  padding: const EdgeInsets.only(left: 0),
+                  child: Icon(
+                    pinned ? Icons.push_pin : Icons.push_pin_outlined,
+                    size: 30,
+                  ),
+                ),
+                title: Text(pinned ? 'Remove Pin' : 'Pin',
                     style: Theme.of(context)
                         .textTheme
                         .bodyText2!
@@ -66,16 +74,37 @@ class ChatListViewModel extends StreamViewModel<QuerySnapshot> {
               ),
               ListTile(
                 onTap: () {
-                  firestoreApi.deleteOnlyRecent(fireuserId);
                   Navigator.pop(context);
+                  showDialog(
+                      context: context,
+                      builder: (context) {
+                        return DuleAlertDialog(
+                          iconBackgroundColor: Colors.red,
+                            title: 'Delete this chat',
+                            icon: FeatherIcons.trash,
+                            primaryButtonText: 'Yes',
+                            primaryOnPressed: () {
+                              firestoreApi.deleteOnlyRecent(fireuserId);
+                              Navigator.pop(context);
+                            },
+                            description: 'This will remove this item from your chat list.',
+                            secondaryButtontext: 'No',
+                            secondaryOnPressed: () {
+                              Navigator.pop(context);
+                            },
+                            );
+                      });
+                  
                 },
-                leading: const Icon(CupertinoIcons.delete_solid),
+                leading: const Icon(FeatherIcons.trash),
                 title: Text('Delete',
                     style: Theme.of(context)
                         .textTheme
                         .bodyText2!
                         .copyWith(fontSize: 18)),
               ),
+              verticalSpaceMedium,
+              bottomPadding(context),
             ],
           ),
         );
